gprs: Extend specification of GSUP cause values (TODO)
Currently, every error message contains a mandantory Cause IE just
referring to GSM 04.08 for further details. This leaves the mapping
from MAP User Errors to GMM Cause values undefined. It is also not
specified, which GMM Cause to assume, if no User Error has been
included in the MAP message.

This commit makes the Cause IE conditional, it is only required if
the MAP message contained a User Error. A mapping from User Errors to
GMM Causes is also added. The default value for GMM Causes in GSUP
error messages is now 'Network failure' to keep the impact low.

TODO:
- the mapping is not finished, it needs review and some values are
  not yet specified.

Sponsored-by: On-Waves ehf